+++ src/main.cpp
@@ -38,7 +38,7 @@
 namespace
 {
 
-constexpr auto version = std::string_view{"0.8.7"};
+constexpr auto version = std::string_view{"0.8.8"};
 
 struct BootstrapConfigResult final
 {